+++ release_notes.md
@@ -4,6 +4,7 @@
   - [#241](https://github.com/Mehdik/Humanizer/pull/241): Added German ToWords localisation
   - [#244](https://github.com/MehdiK/Humanizer/pull/244): Added Slovenian localisation
   - [#247](https://github.com/MehdiK/Humanizer/pull/247): Added Slovenian number to words localisation
+  - [#227](https://github.com/MehdiK/Humanizer/pull/227): Moved factory collections to their own classes, allowed public access via Configurator, and made the default factories lazy loaded
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.24.1...master)
 
@@ -11,7 +12,6 @@
   - [#232](https://github.com/Mehdik/Humanizer/pull/232): Adding code & tests to handle Arabic numbers to ordinal 
   - [#235](https://github.com/Mehdik/Humanizer/pull/235): Fixed the conversion for "1 millon" in SpanishNumberToWordsConverter
   - [#233](https://github.com/Mehdik/Humanizer/pull/233): Added build.cmd and Verify build configuration for strict project build and analysis
-  - [#227](https://github.com/MehdiK/Humanizer/pull/227): Moved factory collections to their own classes, allowed public access via Configurator, and made the default factories lazy loaded
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.23.1...v1.24.1)